Refactor billing code out of DonateViewModel.kt We now have DonationClient.kt to handle connections and donations separately
